feat: Scroll bar jump-to-end buttons (#34)
WLM 2009 featured a jump-to-bottom button seen on the scrollbar in chat
windows, and contained bitmaps for other directions. This commit introduces
the ability to optionally display one of these buttons on any scrollbar.

The buttons are toggled through a simple attached property. Import the
scrollbar namespace `Aerochat.Controls.ScrollBar` into your XAML file (i.e.
as `scrollbar`), and then add a property to a ScrollBar control like
`scrollbar:ScrollBarProperties.ShowJumpToBottom="True"` to display the
button.

The following properties are provided:
- `ShowJumpToTop`
- `ShowJumpToBottom`
- `ShowJumpToLeft`
- `ShowJumpToRight`

Additionally, Chat.xaml has been modified to display the jump-to-bottom
button in the main chat scroll view.

Signed-off-by: Isabella <kawapure@gmail.com>