feat: Add user profile endpoint support to DSharpPlus
This aims to replace getting user profile information through the bot API,
which does not have profile appended to it.

This endpoint does not handle presence, so presence will have to be tracked
in some other way by Aerochat.

The public implementation is currently unfinished. It only provides access
to the DiscordUser object (User property).

Signed-off-by: Isabella <kawapure@gmail.com>