feat: File attachments through the attachments editor
This is currently unpolished, but it works out pretty well I think. As any
meaningful change does, it required a change to the unholy beast that is
DSharpPlus, but fortunately this time it did not require modifying 10,000
files to add one (1) feature.

Please be careful when testing with this as I believe that Discord protects
their Google-provided file-attachment endpoints a bit more than their
regular endpoints.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>