chore: Check dynamic content (news & notices) into main repo
This is the first part of a plan to migrate such dynamic content into the
main repository, which has the added effect that I can manage it. Currently
the sources are owned exclusively by Maddie as part of private GitHub Gists
which I cannot update and which she hasn't updated in over 8 months.

Since users would ideally see new, relevant content here, it just makes way
more sense to store the dynamic information here in the main repository so
that all contributors can edit it.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>